Add and subtract with Fixed
FossilOrigin-Name: 03b7a5233b478133db1a342ecc8cd8320d90b6ee5202b3cc84f9ac30a0d400c3

+++ src/calendar/akan.rs
@@ -108,7 +108,7 @@ impl Epoch for AkanDay {
 
 impl From<FixedDate> for AkanDay {
     fn from(date: FixedDate) -> AkanDay {
-        AkanDay::day_name(i64::from(date) - i64::from(AkanDay::epoch()))
+        AkanDay::day_name(date - AkanDay::epoch())
     }
 }
