+++ src/epoch/unix.rs
@@ -18,18 +18,13 @@ impl EpochMoment for UnixMoment {
 impl TryFrom<UnixMoment> for FixedMoment {
     type Error = CalendarError;
     fn try_from(s: UnixMoment) -> Result<FixedMoment, Self::Error> {
-        FixedMoment::try_from(
-            f64::from(UnixMoment::epoch_moment()) + ((s.0 as f64) / (24.0 * 60.0 * 60.0)),
-        )
+        FixedMoment::try_from(UnixMoment::epoch_moment() + ((s.0 as f64) / (24.0 * 60.0 * 60.0)))
     }
 }
 
 impl From<FixedMoment> for UnixMoment {
     fn from(t: FixedMoment) -> UnixMoment {
-        UnixMoment(
-            (24.0 * 60.0 * 60.0 * (f64::from(t) - f64::from(UnixMoment::epoch_moment()))).round()
-                as i64,
-        )
+        UnixMoment((24.0 * 60.0 * 60.0 * (t - UnixMoment::epoch_moment())).round() as i64)
     }
 }
